tdf#162480: Make sure to use correct anchor point for text box creation
In the ODF import, when importing a table, initially a placeholder 1x1
table is created. When this is done from SwView::InsertMedium, frames
are created for the table and its single cell at that stage. Then the
actual table nodes are created, but frames are not created in parallel,
until the table import is finished.

Importing a text box, it used to be created anchored at the end of the
document, and then the anchor was moved to the correct place.

When a text box was anchored to a cell, the process was like this: the
text content was inserted in the last paragraph outside of the current
table;  and then it was moved to the current cell.  When this was done
from  SwView::InsertMedium,  creation of the text content also created
the frame; then the movement fired client notifications, including the
SwFlyAtContentFrame::SwClientNotify,  which needs the new anchor frame.
With cell other than A1, there was no frames for the new anchor in the
table, and that crashed.

This change inserts the text content into the correct place from start,
which avoids the need to move the anchor later.

Co-authored-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I9dd3a2c5527f3c2dd860244456c617558943453a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171898
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins