chart2: move histogram calcualtion into HistogramChartType
Histogram is special in the way that the input data is not the
data that is being shown in the chart - it needs to be calculated
separately usign the input data. First we put the calculation in
the (Histogram)DataInterpreter but that is not a very good place
to calculate as we don't have the access to the chart properties
and can't influence the calculation with it. This change puts the
calculation into (Histogram)ChartType instead and will trigger
the calculation when the data is requested.

This is the first case of such a thing where we calculate the data
on demand from the input data, so we will probably need further
changes the chart code to better support calculated data sequences.

Change-Id: Ia66f9d245f7d42973c728e1dd9a406b66eb2610a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171952
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>