Make sure to anti-alias fonts on Windows, when required
GetUseFontAAFromSystem was introduced to make sure that texts in generated
raster file match the requested antialiasing selection, not system display
settings (VCL plugins usually render text according to system settings for
screen display, even when lines are not anti-aliased).  The use case there
was disabling antialiasing,  to produce clear output in tests - see commit
e6538f5bdd876911ea30f84a6512c03908e620fd tdf#118966 vcl: add a flag to
determine if AA of fonts is used from the system, 2018-07-28.

But the opposite case is also valid:  bitmap file export with antialiasing
must smooth text,  even when system display antialiasing is disabled. This
has hit testTdf162259 on a particular buildbot,  where the output happened
to be black-and-while (see commit 7a08c89b5af8f98045ad825ace6be2f035ad91af
Workaround a non-antialiased output on one of Windows buildbots, 2024-08-17).
The problem turned out to be RDP settings  used to connect that particular
buildbot,  which happened to disable AA.  This basically means that output
of '--convert-to png' would depend on the system where it's called.

This change extends the effect of GetUseFontAAFromSystem to not only force
disabled text AA, when it returns false and line AA is disabled,  but also
force enabled text AA, when it returns false and line AA is enabled.  This
is a Windows-only change. I will change the test to use this in a separate
commit.

Change-Id: I7071f1bdefb77cbb9156dde3f70feb4cf8be73f3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172031
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins