wina11y: Don't set a11y name that gets overwritten right away
It's unclear to me why an a11y object of a TEXT_FRAME
role with a SCROLL_PANE parent should be treated specially
and use the a11y name of its parent.

But the a11y name set this way is unconditionally overwritten
right away in the lines below anyway.

Therefore, just drop the special handling.

Change-Id: I3cfe63e0882803524f51fe760d38880b30105317
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172122
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins