wina11y: Drop special a11y name handling for paragraph/block quote
The ARIA spec prohibits setting an a11y name for paragraphs [1],
which may have been the reason why an empty name was always
explicitly set in `AccObject::UpdateName`, and that was just
extended to apply for the block quote role when that new role
was added in

    commit f44eae4912e8f2ba34d93a71a609473df976a2cb
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Oct 27 14:53:59 2023 +0200

        tdf#135586 a11y: Add a new BLOCK_QUOTE a11y role

Wile it makes sense to always report an empty a11y
name for paragraphs, ensuring that is not the task of the
platform a11y bridge, but it should be done properly in
the underlying `XAccessibleContext` implementation already,
as this is not specific to Windows/IAccessible2.

At least paragraphs in Writer already report an empty
a11y name. If anything else doesn't, that should be
fixed there.

[1] https://w3c.github.io/aria/#paragraph

Change-Id: I7ffd42acc02fc41cf73174e60c32e8938161c4d3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172123
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins