wina11y: Query new a11y name instead of passing it around
Drop `AccObject::SetName` et al. taking a string param
with the new a11y name to set in favor of using
`AccObject::UpdateName` et al. that updates the a11y name
by querying the current value from the underlying
`XAccessible`.

This makes it unnecessary to pass the new value
around when a `AccessibleEventId::NAME_CHANGED`
event is received, and the new name has to
already be set on the a11y object at the point in
time that the event is received.

Change-Id: I1510bde22ef6d1271f6aeaffe245441cd9e11f19
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172145
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins