wina11y: Drop AccObject::ImplInitializeCreateObj
As the method name and comment suggests, this method
was doing more in the past, but that's no more the case.

The only thing it does is doing an assert, but
there's no need to have a separate method for this,
so just do that right away in the (so far) only caller.

Change-Id: Ief41c8d4ec350622eb043340a4b0bd1079d337a1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172177
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins