wina11y: Consistently assert COM obj creation succeeds
The `AccObject` ctor already has an assert
that creating the COM object succeeded,
so be consistent and drop additional checks still
handling the theoretical case that it might have failed
for some reason. (Running out of memory could be one,
but we don't handle that case in other places that construct
simple objects, e.g. using the `new` expression, either.)

Change-Id: Ic71ca8a256f1e224f0050c2ee9b039d8823e577a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172179
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins