wina11y: Unify handling to set action description
Drop `AccObject::UpdateDefaultAction` and let
`AccObject::UpdateActionDesc` take care of
setting the action description for all roles.

It's unclear to me

* why the action description would only be set
  for particular roles

* why `AccObject::UpdateActionDesc` would call
  `AccObject::UpdateDefaultAction` for particular
  roles if it was basically setting the description
  the same way that `AccObject::UpdateActionDesc`
  does directly for other roles (except for
  `AccObject::UpdateDefaultAction` not checking
  that there's at least one action).

Also simplify the code a bit while at it.

Testing with the "Don't Save" button in the "Save Document?"
dialog seen when pressing Alt+F4 with a modified Writer doc
gives the expected result in NVDA's Python console:

    >>> focus.role
    <Role.BUTTON: 9>
    >>> focus.name
    'Don’t Save'
    >>> focus.IAccessibleActionObject.nActions()
    1
    >>> focus.IAccessibleActionObject.description(0)
    'press'

If any particular reason for not setting the action description
for particular roles should show up in the future, they can
still be excluded in one central place now.

Change-Id: I5569f70739135172ddb604f1f1bd958e32626ab0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172182
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>