wina11y: Retrieve default action description on demand
Instead of doing extra work to keep
`CMAccessible::m_pszActionDescription` up to date,
move the logic to retrieve the action description
from the underlying `XAccessibleAction` into
`CMAccessible::get_accDefaultAction` and do that
on demand.

This is similar to what was done for the accessible
description in this commit:

    commit fcf4a26275d7503835f9aa23cb94938809840300
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jan 5 13:41:53 2022 +0000

        tdf#146306 wina11y: Retrieve accessible desc on demand

As expected, the example mentioned in the commit
message of

    Change-Id: I5569f70739135172ddb604f1f1bd958e32626ab0
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Aug 21 09:49:51 2024 +0100

        wina11y: Unify handling to set action description

still behaves the same as without this commit in place.

Change-Id: I7d1ea0ce302df955314c8e46c530cc24d30fcd4b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172183
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins