Make rendering more flexible, render fields in separate layer
If the amster page has a field (like slide number) then render the
objects containing fields into a separate master slide layer.

Also change the rendering in such a way to be more flexible and
has less conditions. So in the renderer it doesn't really matter
if we are rendering a master slide or slide, what matters only is
if the object is already rendered and if we need to stop rendering
for some specified reason (like we encountered a field and need to
stop rendering and need to switch to a new layer).

Change-Id: I37ea2528427bbc1b3de938f960fb344866ee9399
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/176911
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>