wina11y: Only require action iface for the actual object
Quoting from the `IAccessible::accDoDefaultAction` doc [1]
for the `varChild` param:

> Specifies whether the default action belongs to the object
> or one of the object's child elements.

If the default object of a child and not the object itself
should be triggered, there's no need for the object
itself to also support actions itself.

Therefore, only perform such a check in the
`CHILDID_SELF` case.

Also, return `S_FALSE` in case there's no action, as this is
not a case of "An unknown or generic error occurred.", which
is the meaning of `E_FAIL` [2].

as other places do
move that check in `CMAccessible::get_accDefaultAction`, if a child
object is specified instead of requesting

[1] https://learn.microsoft.com/en-us/windows/win32/api/oleacc/nf-oleacc-iaccessible-accdodefaultaction
[2] https://learn.microsoft.com/en-us/windows/win32/winauto/return-values

Change-Id: Idcd3ab8a415e73e496cf027506428518a9f4a203
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172184
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins