wina11y: Query XAccessibleAction iface on demand
Despite the

    // initially m_xAction and m_xContext are the same object
    // but they may be different once AccObject::UpdateAction() is called?

comment in winaccessibility/source/UAccCOM/MAccessible.h,
`MAccessible::m_xAction` is supposed to point to the same
underlying object as `MAccessible::m_xContext`, see
`AccObject::UpdateAction` (which queries for the
`XAccessibleAction` interface and sets that).

Simplify this by just querying `MAccessible::m_xContext`
for the `XAccessibleAction` iface on demand in those
places where it's needed.

Change-Id: Id3d89213937c2c72a936947da69ca7fddc8f1b09
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172205
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>