tdf#157637 ODT->DOC: use layoutInCell for para-oriented fly
This is a very sticky business.
LO and Word implement aspects of FollowTextFlow so differently.

But for native LO documents, it is clear that at least
the image placement of paragraph-oriented flies
is based on the cell paragraph, which in MSO layout
means the layoutInCell paragraph.

Prior to 7.4, we never wrote anything related to layoutInCell,
and since the default is true, layoutInCell was always true,
so writing layoutInCell = false less often can't cause a regression.

But of course there are implications.
The main side effect is that the cell will expand to contain the fly,
or the UI might, on an "OK", move or squish the image into the cell.
But we already had that prior to 7.4 anyway.

WrapThrough shouldn't come into play here
(although they tend to look better, with less side effects).

make CppunitTest_sw_ww8export3 \
    CPPUNIT_TEST_NAME=testTdf79186_noLayoutInCell

----
This cannot fix LO inserting images into a doc-formatted file though.
That can only be handled by "fixing" the layout to orient the newly
inserted image relative to the table-paragraph
instead of the cell-paragraph.

LO currently does not properly reflect the layout
when layoutInCell=false for DOC/DOCX<=compat14 flies,
so simply looking at LO results is very deceiving.
Make sure you also check how the end result looks in MS Word.

Change-Id: Ia631c56a25ad103ac448f1f46a8553ab1478330b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/157662
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins