tdf#91739 wina11y: Ensure IAccessible for relation target exists
When retrieving the `IAccessible` for the relation
target's `XAccessible` and
`CMAccessible::get_IAccessibleFromXAccessible` initially
returns nullptr, insert the a corresponding object
and retrieve it then.

The same is already done in most places calling
`CMAccessible::get_IAccessibleFromXAccessible`, e.g.
`CAccTableCell::get_columnHeaderCells` or
`CEnumVariant::Next`.

This makes sure that a proper `IAccessible` is
returned instead of triggering an error if no
such object has been created yet.

This is the case e.g. for off-screen paragraph
objects in Writer that can still be accessed
on the UNO a11y level via the
`AccessibleRelationType::CONTENT_FLOWS_TO` relation.
However, when testing with the test docs from tdf#91739,
the issue even occured much earlier, with on-screen
paragraphs already that weren't available via the
FLOWS_TO relation from their preceding paragraphs.

With this commit in place, having NVDA read out
the whole sample documents using the NVDA + Down shortcut
works now for the tdf#91739 sample docs, in the same way
it already does for Orca on Linux.

For more details, see also the previous analysis in
NVDA ticket [1].

[1] https://github.com/nvaccess/nvda/issues/4115#issuecomment-2299084993

Change-Id: I2ef06c6efa368f10dcf5d917d781dabc8ce9a65f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172215
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>