wina11y: Explicitly map UNO to IA2 relation type
Use switch/case to explicitly map from UNO's
`AccessibleRelationType` to the IAccessible2 relation
type strings, instead of implicitly using the integer
values of the `AccessibleRelationType` as indices into
a map.

Let the caller allocate the string to return.

Use `sal_Int16` for the `AccessibleRelationType` param
to match the type of the actual integer values.

Change-Id: I8c094d763ca602131a578f20719d00cad53088d0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172257
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>