tdf#91739 wina11y: Implement IAccessible2_2::get_relationTargetsOfType
Implement the `IAccessible2_2::get_relationTargetsOfType`
method [1] in the Windows accessibility bridge.

As described in NVDA issue comment [2], this method is
the preferred way that NVDA uses to get e.g. the target
object of the "flowsTo" relation.

With this commit in place, retrieving the next Writer
paragraph in the NVDA Python console using that relation
now also works with the fallback code path in NVDA [3]
dropped in a local build of NVDA:

    >>> focus.flowsTo
    <NVDAObjects.Dynamic_SymphonyParagraphSymphonyTextEditableTextWithAutoSelectDetectionIAccessible object at 0x0B8F5890>

[1] https://accessibility.linuxfoundation.org/a11yspecs/ia2/docs/html/interface_i_accessible2__2.html#a63f214b322c663caf01d4bb67277f5af
[2] https://github.com/nvaccess/nvda/issues/4115#issuecomment-2299084993
[3] https://github.com/nvaccess/nvda/blob/df6bd0770db12ff73905251054d12f01911030dd/source/NVDAObjects/IAccessible/__init__.py#L1861-L1874

Change-Id: Ifcd15527b4a3f5302168942f19248a0ba9a4d306
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172258
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>