tdf#162541 doc import NOT-layoutInCell: hori-para => PAGE_PRINT_AREA
If the fly is not layoutInCell,
then MSO applies the para-orientation
against the paragraph that contains the entire table,
not the cell paragraph it is anchored to.

For the horizontal case, this table-paragraph is equivalent
to the page margin, so change it to a meaningful value.

The vertical case cannot be meaningfully mapped to an equivalent,
so someone will need to implement that in layout code.

Wrap through has NOTHING to do with this.

make CppunitTest_sw_ww8export4 CPPUNIT_TEST_NAME=testTdf162541

Change-Id: I66212a9cda89e63e7b17daba0661a90f681fc0e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172218
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>