tdf#162541 docx import NOT-layoutInCell: hori-para => PAGE_PRINT_AREA
This was copied from DOC import.
Since compat15 is always layoutInCell, the impact for DOCX is minimal.

If the fly is not layoutInCell,
then MSO applies the para-orientation
against the paragraph that contains the entire table,
not the cell paragraph it is anchored to.

For the horizontal case, this table-paragraph is equivalent
to the page margin, so change it to a meaningful value.

The vertical case cannot be meaningfully mapped to an equivalent,
so someone will need to implement that in layout code.

Wrap through has NOTHING to do with this.

make CppunitTest_sw_ooxmlexport21 CPPUNIT_TEST_NAME=testTdf162541

Change-Id: I17b28f9d77cd85b4104e8c8f0c347e61b6c8f5a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172297
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>