Resolves: tdf#162597 'ok' in chart range dialog doesn't commit datarange
so nothing happens.

since:

commit 4965b055b5998dc345ecaeb80fc8c3874b19e2a5
CommitDate: Fri Feb 16 12:08:55 2024 +0100

    Chart: Make Data Range Dialog Async

async conversion missed that there is a DataSourceDialog::run
override so the use of asyncRun means the body of that isn't run.

To keep things simple we can add an explicit 'ok' handler and move
the special dialog ok code into that instead, so all the existing
asyncRun's used with this dialog run the same code originally called
in the previous non-async case.

Change-Id: Ibdb4b764b1a1e48c01ee1528bc260f694e840221
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172336
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>