tdf#162614: report the reason why a frame couldn't grow
... and use it in SwTextFrameBreak::IsBreakNow to decide if the content
needs to go to a follow or not.

The problem was, that in a specific case of the bugdoc, the follow text
line was taller than the fixed row height on that page; that resulted
in an attempt to break it again, creating an empty follow in the same
table cell; it was not moved to the next page, detecting the problem in
SwTextFrame::CalcFollow, and the formatting looped, until loop control
fired, calling SwLooping::Drastic, which validated all page's content,
including tables below, that had invalid zero client (prn) height.

It was very difficult to tell when the break is needed, and when it's
not, without the information why some frame eventually denied growing
its height by the requested amount. In the bugdoc case, I failed to
find at the SwTextFrameBreak::IsBreakNow level anything that could be
used as a reliable distinction; some heuristic conditions used there
broke layout in unit tests left and right. So eventually, I came with
this mechanism of reporting the "why I can't grow" reason back to the
caller, using the new out argument.

I believe that, beyond this use case, it might be useful in many other
cases, both to ease the decision, and to prevent needless trial-and-
error iterations, increasing stability and performance. The enum used
for the reason reporting includes 'FlowToFollow' and 'BalancedColumns',
that could be used in those other cases. However, for my needs here, I
only need the FixedSizeFrame reason, telling that it's not a case when
breaking to a next frame is possible, so preventing the break.

The choice of reported reasons in this patch is mostly best guess; it
may be wrong in some places. The crucial for the current fix is the
assignment inside the '!(GetType() & nTmpType) && HasFixSize()' check
of SwLayoutFrame::GrowFrame.

Change-Id: I9ce5dd4d2298b60e186fdf485efb85ab304308ee
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172362
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins