tdf#162612 mso import NOT-layoutInCell: effectively are wrap-through
The way Microsoft implements the layout of in-table-anchored shapes
that are NOT marked as layoutInCell
is to treat them all as wrap-through.

So, lets explicitly mark them as they ought to be during import.

tdf154863-img-move-crash.docx is a fine example of this,
although the document itself is ugly and too complex for a unit test.

make CppunitTest_sw_ooxmlexport21 CPPUNIT_TEST_NAME=testTdf162612
make CppunitTest_sw_ww8export4 CPPUNIT_TEST_NAME=testTdf162541

There is a SLIGHT danger here, as shown by testTdf162541.
LO looks VERY different from Microsoft's VERY WEIRD LAYOUT,
but will "correct" it with a round-trip.
I'm not too concerned - LO is exporting the document
the way it sees it.

Change-Id: I14d8ea9dcef257a3239a9601b63c04cde780cb33
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172364
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins