Make CPDB work with --enable-cpdb --disable-cups
The `PrinterInfoManager` implementation in
vcl/unx/generic/printer/printerinfomanager.cxx
takes care of actually using the `CUPSManager`
or the `CPDManager` for handling printers/printing.

Therefore, build that one also when `--disable-cups`
is used in addition to `--enable-cpdb`, and only make
the CUPS-specific bits depend on `ENABLE_CUPS`,
rather than building the dummy version of the
`PrinterInfoManager` implemented in
vcl/null/printerinfomanager.cxx.

This makes the CPDB printers show up in the print
dialog with WIP Gerrit change [1] in place in
addition.

[1] https://gerrit.libreoffice.org/c/core/+/169617/25

Change-Id: I5dddc72b8e679a84225f1dfb9c87d7562613a9f7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172401
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins