text drawing boxes don't redraw with a new text autocolor choice
when light/dark mode themes are toggled on off in calc. This works fine
in writer though, so follow the same pattern as SwViewShellImp::PaintLayer
and set a default background color on the Outliner when we draw.

So SdrTextPrimitive2D::get2DDecomposition will see that the outliner
should have a new background, and drops its cache, even if
setSuitableOutlinerBg has no reason to set anything explicit.

Change-Id: I19ca407853238e6533085c6aa617206ee122d3a9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172425
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>