vcl: fix testTdf129085
For unclear reasons, after upgrading to Fedora 40 and pulling the latest
master, testTdf129085 started to fail:

 pdfexport2.cxx:328:Assertion
 Test name: (anonymous namespace)::testTdf129085::TestBody
 equality assertion failed
 - Expected: 0x0(Error Area:Io Class:NONE Code:0)
 - Actual  : 0x8203(Error Area:Vcl Class:General Code:3)

The test runs in the libreoffice-24-2 branch, and the difference is that
in the old brach a JPEG was exported into the PDF as expected, but in
master a DIB was exported instead, because it turned out to be smaller
than the JPEG, because the quality level of the JPEG was 90 in master
and 80 in the old branch, due to the old branch test passing FilterData
with UseTaggedPDF, which was later removed from the test due to being
the default.

Pass in an explicit low quality to prevent this problem.

Change-Id: Ic26c25b15fecd52c5195a7d30c603bda34fe88d3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172470
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>