tdf#161805 cui: Populate line tab menus when page gets created
Since

    commit f075fa01cb4f74185f13eb0a8d7f84cf1f47af49
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Aug 22 10:26:32 2023 +0200

        tdf#141101 tdf#101886 a11y: Restore previous focus on col/line popup close

, the menu button toggled handler is no more called before
the menu shows.

`SvxLineTabPage` was relying on the old behavior and
only populating the menu entries in the handler,
resulting in them not being present any more the
first time the menu gets opened after the above
commit.

Populate the menu items when the tab page gets created
instead, by changing the handler to a regular method
`SvxLineTabPage::PopulateMenus` that gets called
from `SvxLineTabPage::Construct` (which already takes
care of updating the list boxes in the tab page).

Move the call to `SvxLineTabPage::Construct` to the
end of `SvxLineTabPage::PageCreated`, as
`SvxLineTabPage::PopulateMenus` makes use of
`m_pSymbolList` and `m_xSymbolAttr` that were previously
only set after the call to `SvxLineTabPage::Construct`.

Change-Id: I521a50f3a3695f20f789add89ffd88749e3b5cd2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172472
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>