Refactor impl_packLocalizedProperties
Make it more C++ish.
The change made two functional modifications:

1. It now doesn't filter out non-string values of the localized
properties. Since commit a5649055a0ec54ded05a879d018903628bfe2f0b
(INTEGRATION: CWS xmlfilter01 (1.38.18); FILE MERGED, 2003-04-04),
'OUString sLocaleValue' intermediate local variable was used for
values taken from XNameAccess::getByName before storing it to the
Sequence<PropertyValue>. It contradicts the documentation of the
impl_packLocalizedProperties function, created in commit
4bbc40bf5c29606c44934b18ed5962db8a0bc5d1 (#84766# new mode for
locale=*, 2001-04-11), which explicitly tells that the value type
"depends from cfg entry". The commit message of the later commit
is not clear, mentioning some "repair ALLO_LOCALE mode"; but it
likely refers to the change of CONFIG_MODE_ALL_LOCALES check that
was made bitwise, instead of direct ==. I think that sLocaleValue
was used during debugging, to check the value, similar to another
debug variable - sLocaleName - also introduced in that commit; it
was simply erroneously kept in the final patch.

2. When lOutValues has an XInterface, but not XNameContainer, it
now is copied to the out sequence (it was dropped before). Likely
it won't really happen in practice, but the limitation reason was
unclear.

The use of a sequence declared out of the loop, and realloc'd in
the loop, is in fact pessimization, because such a realloc would
need to copy the existing values, only to discard them later.

Change-Id: I726c7dbeae58cbf92b53233ae4eeb3cd070375ec
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172507
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins